+++ homepage/homepage/content/docs/code-snippets/core-concepts/sync-and-storage/vanilla.ts
@@ -1,11 +1,9 @@
-import { createJazzBrowserContext } from "jazz-tools/browser";
+import { JazzBrowserContextManager } from 'jazz-tools/browser';
 
-// Get a free API Key at dashboard.jazz.tools, or use your email as a temporary key.
+// [!code hide:1]
 const apiKey = "you@example.com";
-// @ts-expect-error only showing sync options
-const ctx = await createJazzBrowserContext({
+await new JazzBrowserContextManager().createContext({
   sync: {
-    peer: `wss://cloud.jazz.tools/?key=${apiKey}`,
+    peer: `wss://cloud.jazz.tools?key=${apiKey}`,
   },
-  // Rest of config
 });