+++ packages/cojson/src/storage/knownState.ts
@@ -11,14 +11,14 @@ import {
  * and provides the API to wait for the data to be fully stored.
  */
 export class StorageKnownState {
-  knwonStates = new Map<string, CoValueKnownState>();
+  knownStates = new Map<string, CoValueKnownState>();
 
   getKnownState(id: string): CoValueKnownState {
-    const knownState = this.knwonStates.get(id);
+    const knownState = this.knownStates.get(id);
 
     if (!knownState) {
       const empty = emptyKnownState(id as RawCoID);
-      this.knwonStates.set(id, empty);
+      this.knownStates.set(id, empty);
       return empty;
     }
 
@@ -26,7 +26,7 @@ export class StorageKnownState {
   }
 
   setKnownState(id: string, knownState: CoValueKnownState) {
-    this.knwonStates.set(id, knownState);
+    this.knownStates.set(id, knownState);
   }
 
   handleUpdate(id: string, knownState: CoValueKnownState) {