+++ packages/cojson-transport-ws/CHANGELOG.md
@@ -1,5 +1,18 @@
 # cojson-transport-nodejs-ws
 
+## 0.19.19
+
+### Patch Changes
+
+- Updated dependencies [83f84ca]
+  - cojson@0.19.19
+
+## 0.19.18
+
+### Patch Changes
+
+- cojson@0.19.18
+
 ## 0.19.17
 
 ### Patch Changes