+++ packages/jazz-webhook/CHANGELOG.md
@@ -1,5 +1,29 @@
 # jazz-webhook
 
+## 0.19.19
+
+### Patch Changes
+
+- Updated dependencies [171e1c6]
+- Updated dependencies [053a283]
+- Updated dependencies [41b2cf4]
+- Updated dependencies [923bc8e]
+- Updated dependencies [83f84ca]
+  - jazz-tools@0.19.19
+  - cojson@0.19.19
+  - cojson-storage-sqlite@0.19.19
+  - cojson-transport-ws@0.19.19
+
+## 0.19.18
+
+### Patch Changes
+
+- Updated dependencies [729d46c]
+  - jazz-tools@0.19.18
+  - cojson@0.19.18
+  - cojson-storage-sqlite@0.19.18
+  - cojson-transport-ws@0.19.18
+
 ## 0.19.17
 
 ### Patch Changes